update plugin focus
