+++ docs/recipes/plugin/lighthouse-plugin-preload-as/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "lighthouse-plugin-preload-as",
+  "private": true,
+  "main": "./plugin-config.js"
+}
