+++ docs/recipes/plugin/lighthouse-plugin-preload-as/plugin-config.js
